userId in Dto is deleted if need userId for authorizion using CoxtHolder

+++ src/main/java/com/vaccinelife/vaccinelifeapi/service/CommentService.java
@@ -8,8 +8,12 @@ import com.vaccinelife.vaccinelifeapi.repository.CommentRepository;
 import com.vaccinelife.vaccinelifeapi.repository.UserRepository;
 import com.vaccinelife.vaccinelifeapi.repository.VacBoardRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.client.HttpClientErrorException;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -24,17 +28,15 @@ public class CommentService {
 
 
     @Transactional
-    public void createComment(CommentPostRequestDto requestDto) {
+    public void createComment(CommentPostRequestDto requestDto, Long userId) {
         VacBoard vacBoard = vacBoardRepository.findById(requestDto.getVacBoardId()).orElseThrow(
                 () -> new IllegalArgumentException("해당 게시물이 존재하지 않습니다.")
         );
-        User user = userRepository.findById(requestDto.getUserId()).orElseThrow(
-                () -> new IllegalArgumentException("해당 아이디가 존재하지 안습니다.")
-        );
+        User user = userRepository.getById(userId);
 
         List<Comment> commentCount = commentRepository.findByVacBoardId(requestDto.getVacBoardId());
         int commentSize = commentCount.size();
-        vacBoard.setCommentCount(commentSize+1);
+        vacBoard.setCommentCount(commentSize + 1);
 
         Comment comment = Comment.builder()
                 .user(user)
@@ -50,22 +52,31 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteComment( Long vacBoardId,Long id) {
+    public void deleteComment(Long id) {
 
         Comment comment = commentRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("해당 댓글을 찾을 수 없습니다.")
         );
-        VacBoard vacBoard = vacBoardRepository.findById(vacBoardId).orElseThrow(
-                () -> new IllegalArgumentException("해당 게시물을 찾을 수 없습니다.")
+        Long vacBoardId = comment.getVacBoard().getId();
+        VacBoard vacBoard = comment.getVacBoard();
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Object principal = authentication.getPrincipal();
+        String username = principal.toString();
+
+        User user = userRepository.findByUsername(username).orElseThrow(
+                () -> new IllegalArgumentException("해당 유저는 존재 하지 않습니다.")
         );
-
-        List<Comment> commentCount = commentRepository.findByVacBoardId(vacBoardId);
-        int commentSize = commentCount.size();
-        vacBoard.setCommentCount(commentSize-1);
-
-        commentRepository.delete(comment);
-
-
+        Long realId = user.getId();
+        Long anonymousId = comment.getUser().getId();
+        if (realId.equals(anonymousId)) {
+            List<Comment> commentCount = commentRepository.findByVacBoardId(vacBoardId);
+            int commentSize = commentCount.size();
+            vacBoard.setCommentCount(commentSize - 1);
+            commentRepository.delete(comment);
+        } else {
+            throw new IllegalArgumentException("본인이 작성한 댓글만 삭제할 수 있습니다.");
+        }
     }
 
 }
